move and rename script files
