+++ src/price_indexr/__init__.py
@@ -1,0 +1,4 @@
+from price_indexr.collect import (
+    collect_prices,
+    
+)
